pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "modeling-tools"
version = "2.0.0"
description = "Model visualization and simulation tool"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # NEEDED BY APP
    "PyQt6",
    "matplotlib",
    "numpy",
    "scienceplots",
    "pyyaml",
    "platformdirs",
    # NEEDED BY SPECIFIC MODELS
    "scipy",
    "mesa",
    "networkx"
]

[project.scripts]
modeling-tools = "modeling_tools.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
